+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * `sjp.lmm`  and `sjp.glmm` now also support linear mixed effects models (of class `merMod`).
 * `sjp.int` now uses proper x-axis-tick-labels for `type = "eff"`, when predictor on x-axis is a factor with non-numeric factor-levels (or has label attributes).
+* `sjp.glm` gets a `group.estimates` argument to group estimates in forest plots and colour them according to group assignment. Use arguments `show.legend` and `legendTitle` to modify group legend.
 
 
 ## Bug fixes